recut(M2): upstream re-cut adaptations for the M2 files
Carried from the re-cut integrate 228410e4 (adaptations onto upstream main 202be09d), the cap split 8c2311a3 (test-only files split, GPU-qualified production files allow-listed with rationale) and the workspace rustfmt/clippy pass of session 01 (2026-09-06): the per-file residual between the replayed candidate commits and the frozen tree d5b44da1, each hunk placed in the unit whose commits introduced the lines it touches (git blame). Files: .github/workflows/file-size-cap.yml crates/atlas-core/src/config/tests.rs crates/spark-model/src/layers/dense_ffn_kernel_tests.rs crates/spark-model/src/layers/dense_ffn_native_batch_tests.rs crates/spark-model/src/layers/glm5next_mlp/build.rs crates/spark-model/src/model/impl_a3.rs crates/spark-model/src/model/trait_impl/decode_a.rs crates/spark-runtime/src/buffers/accessors.rs crates/spark-runtime/src/cuda_backend.rs crates/spark-runtime/src/cuda_backend/gpu_impl.rs crates/spark-runtime/src/weights.rs

+++ crates/spark-model/src/layers/glm5next_mlp/build.rs
@@ -131,8 +131,8 @@ fn build_expert_ptr_table(
             continue; // remote — null stays, and the kernel skips the slot
         };
         let p = proj(&experts[local]);
-        packed[id * 8..id * 8 + 8].copy_from_slice(&(p.packed.0 as u64).to_le_bytes());
-        scale[id * 8..id * 8 + 8].copy_from_slice(&(p.scale.0 as u64).to_le_bytes());
+        packed[id * 8..id * 8 + 8].copy_from_slice(&p.packed.0.to_le_bytes());
+        scale[id * 8..id * 8 + 8].copy_from_slice(&p.scale.0.to_le_bytes());
         scale2[id * 4..id * 4 + 4].copy_from_slice(&p.scale_2.to_le_bytes());
     }
     let packed_ptrs = gpu.alloc(packed.len())?;
